chore: moved usage of compass-preferences-model to compass-preferences-model/provider
