+++ packages/compass-aggregations/src/components/pipeline-toolbar/pipeline-header/index.tsx
@@ -14,7 +14,7 @@ import PipelineStages from './pipeline-stages';
 import PipelineActions from './pipeline-actions';
 import SavedPipelines from '../../saved-pipelines/saved-pipelines';
 import type { RootState } from '../../../modules';
-import { usePreference } from 'compass-preferences-model';
+import { usePreference } from 'compass-preferences-model/provider';
 
 const containerStyles = css({
   display: 'flex',
