+++ packages/compass-aggregations/src/components/pipeline-toolbar/pipeline-header/pipeline-stages.tsx
@@ -11,7 +11,7 @@ import {
   Icon,
 } from '@mongodb-js/compass-components';
 import { AIExperienceEntry } from '@mongodb-js/compass-generative-ai';
-import { useIsAIFeatureEnabled } from 'compass-preferences-model';
+import { useIsAIFeatureEnabled } from 'compass-preferences-model/provider';
 
 import type { RootState } from '../../../modules';
 import { editPipeline } from '../../../modules/workspace';
