+++ packages/compass-aggregations/src/components/pipeline-toolbar/pipeline-options/pipeline-collation.tsx
@@ -14,7 +14,7 @@ import type { RootState } from '../../../modules';
 import { collationStringChanged } from '../../../modules/collation-string';
 import { maxTimeMSChanged } from '../../../modules/max-time-ms';
 import { DEFAULT_MAX_TIME_MS } from '../../../constants';
-import { usePreference } from 'compass-preferences-model';
+import { usePreference } from 'compass-preferences-model/provider';
 
 const pipelineOptionsContainerStyles = css({
   paddingTop: spacing[1],
