+++ packages/compass-aggregations/src/components/stage-preview/output-stage-preview.tsx
@@ -20,7 +20,7 @@ import {
   MERGE_STAGE_PREVIEW_TEXT,
   OUT_STAGE_PREVIEW_TEXT,
 } from '../../constants';
-import { usePreference } from 'compass-preferences-model';
+import { usePreference } from 'compass-preferences-model/provider';
 
 const stagePreviewStyles = css({
   display: 'flex',
