+++ packages/compass-aggregations/src/components/stage-toolbar/index.tsx
@@ -20,7 +20,7 @@ import { enableFocusMode } from '../../modules/focus-mode';
 import OptionMenu from './option-menu';
 import type { StoreStage } from '../../modules/pipeline-builder/stage-editor';
 import { getInsightForStage } from '../../utils/insights';
-import { usePreference } from 'compass-preferences-model';
+import { usePreference } from 'compass-preferences-model/provider';
 import type { ServerEnvironment } from '../../modules/env';
 import { createSearchIndex } from '../../modules/search-indexes';
