+++ packages/compass-aggregations/src/components/stage-toolbar/stage-operator-select.tsx
@@ -1,5 +1,5 @@
 import React, { useCallback } from 'react';
-import { withPreferences } from 'compass-preferences-model';
+import { withPreferences } from 'compass-preferences-model/provider';
 import { connect } from 'react-redux';
 
 import {
