+++ packages/compass-aggregations/src/plugin.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Aggregations from './components/aggregations';
 import { ConfirmationModalArea } from '@mongodb-js/compass-components';
-import { usePreference } from 'compass-preferences-model';
+import { usePreference } from 'compass-preferences-model/provider';
 import type { ConfigureStoreOptions } from './stores/store';
 
 export const AggregationsPlugin: React.FunctionComponent<
