+++ packages/compass-crud/src/components/update-data-menu.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Icon, Button, Tooltip } from '@mongodb-js/compass-components';
-import { usePreference } from 'compass-preferences-model';
+import { usePreference } from 'compass-preferences-model/provider';
 
 type UpdateMenuButtonProps = {
   isWritable: boolean;
