+++ packages/compass-databases-navigation/src/collection-item.tsx
@@ -21,7 +21,7 @@ import type {
   NamespaceItemProps,
 } from './tree-item';
 import type { Actions } from './constants';
-import { usePreference } from 'compass-preferences-model';
+import { usePreference } from 'compass-preferences-model/provider';
 
 const CollectionIcon: React.FunctionComponent<{
   type: string;
