+++ packages/compass-explain-plan/src/components/explain-plan-side-summary.tsx
@@ -20,7 +20,7 @@ import {
   type SerializedExplainPlan,
 } from '../stores/explain-plan-modal-store';
 import { connect } from 'react-redux';
-import { usePreference } from 'compass-preferences-model';
+import { usePreference } from 'compass-preferences-model/provider';
 
 const defaultFormatter = (_: unknown) => String(_);
