+++ packages/compass-indexes/src/components/regular-indexes-table/regular-indexes-table.tsx
@@ -1,7 +1,7 @@
 import React, { useMemo } from 'react';
 import { connect } from 'react-redux';
 import { IndexKeysBadge } from '@mongodb-js/compass-components';
-import { withPreferences } from 'compass-preferences-model';
+import { withPreferences } from 'compass-preferences-model/provider';
 
 import type { RootState } from '../../modules';
