+++ packages/compass-preferences-model/src/utils.ts
@@ -1,4 +1,4 @@
-import { usePreference } from '.';
+import { usePreference } from './react';
 import type {
   AllPreferences,
   ParsedGlobalPreferencesResult,
